# docs/api-gateway/swagger.yaml
basePath: /
definitions:
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.HotelRequest:
    properties:
      address:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.HotelResponse:
    properties:
      address:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      room_types:
        items:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomTypeSummary'
        type: array
    type: object
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentRequest:
    properties:
      amount:
        type: number
      booking_id:
        type: string
      currency:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentResponse:
    properties:
      id:
        type: string
      payment_url:
        type: string
      provider:
        type: string
      status:
        type: string
    type: object
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.RefundRequest:
    properties:
      amount:
        type: number
      payment_id:
        type: string
      reason:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.RefundResponse:
    properties:
      id:
        type: string
      reference:
        type: string
      status:
        type: string
    type: object
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomRequest:
    properties:
      number:
        type: string
      room_type_id:
        type: string
      status:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomTypeRequest:
    properties:
      amenities:
        type: string
      base_price:
        type: number
      capacity:
        type: integer
      hotel_id:
        type: string
      name:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomTypeSummary:
    properties:
      capacity:
        type: integer
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
info:
  contact: {}
  description: Aggregated API surface for hotel booking platform.
  title: Hotel Booking Microservices API
  version: "1.0"
paths:
...
  /hotels:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.HotelResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      summary: List hotels
      tags:
      - Hotels
    post:
      consumes:
      - application/json
      parameters:
      - description: Hotel payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.HotelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create hotel
      tags:
      - Hotels
...
  /payments:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payment payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Initiate payment
      tags:
      - Payments
  /payments/{id}:
    get:
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payment
      tags:
      - Payments
  /payments/refund:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refund payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RefundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RefundResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Refund payment
      tags:
      - Payments
  /payments/webhook:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      summary: Payment webhook
      tags:
      - Payments
  /room-types:
    post:
      consumes:
      - application/json
      parameters:
      - description: Room type payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create room type
      tags:
      - Hotels
  /rooms:
    post:
      consumes:
      - application/json
      parameters:
      - description: Room payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create room
      tags:
      - Hotels
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"